Исправил некоторые замечания и поработал немного с контроллером
