+++ src/components/edit-event/edit-event.js
@@ -1,0 +1,8 @@
+import AbstractComponent from "../abstract-component";
+import {editEventTemplate} from "./edit-event-tpl";
+
+export default class EditedEvent extends AbstractComponent {
+  getTemplate() {
+    return editEventTemplate(this._eventItem);
+  }
+}
