--- src/components/menu.js
@@ -1,34 +1,0 @@
-import {createElement} from "../utils.js";
-
-const createMenuTemplate = () => {
-  return (
-    `
-      <nav class="trip-controls__trip-tabs  trip-tabs">
-      <a class="trip-tabs__btn" href="#">Table</a>
-      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
-      </nav>
-      `
-  );
-};
-
-export default class Menu {
-  constructor() {
-    this._element = null;
-  }
-
-  getTemplate() {
-    return createMenuTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeelement() {
-    this._element = null;
-  }
-}
