+++ src/components/menu/menu-tpl.js
@@ -1,0 +1,10 @@
+export const createMenuTemplate = () => {
+  return (
+    `
+        <nav class="trip-controls__trip-tabs  trip-tabs">
+        <a class="trip-tabs__btn" href="#">Table</a>
+        <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
+        </nav>
+        `
+  );
+};
