+++ src/components/sort-form/sort-form.js
@@ -1,0 +1,9 @@
+import AbstractComponent from "../abstract-component";
+import {createSortFormTemplate} from "./sort-form-tpl";
+
+export default class SortForm extends AbstractComponent {
+  getTemplate() {
+    return createSortFormTemplate();
+  }
+}
+
