+++ src/components/trip-info/trip-info.js
@@ -1,0 +1,8 @@
+import AbstractComponent from "../abstract-component";
+import {createTripInfoTemplate} from "./trip-info-tpl";
+
+export default class TripInfo extends AbstractComponent {
+  getTemplate() {
+    return createTripInfoTemplate();
+  }
+}
