--- src/utils.js
@@ -1,31 +1,0 @@
-import {RenderPosition} from "./mock/data.js";
-
-export const renderTemplate = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.append(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement;
-};
-
-export const getRandomBetween = (min, max) => {
-  return min + Math.floor(Math.random() * (max - min));
-};
-
-export const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomBetween(0, array.length);
-
-  return array[randomIndex];
-};
