+++ src/utils/common.js
@@ -1,0 +1,10 @@
+
+export const getRandomBetween = (min, max) => {
+  return min + Math.floor(Math.random() * (max - min));
+};
+
+export const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomBetween(0, array.length);
+
+  return array[randomIndex];
+};
